+++ paradigm.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python3
 # paradigm.py -- generate paradigms
-# Andreas Nolda 2022-08-22
+# Andreas Nolda 2022-08-23
 
 import sys
 import os
@@ -12,13 +12,13 @@ from dwdsmor import analyse_word
 from blessings import Terminal
 from collections import namedtuple
 
-version = 4.4
+version = 5.0
 
 basedir = os.path.dirname(__file__)
 libdir  = os.path.join(basedir, "lib")
 libfile = os.path.join(libdir, "dwdsmor-index.a")
 
-indices     = [1, 2, 3, 4]
+indices     = [1, 2, 3, 4, 5]
 psos        = ["ADJ", "ART", "CARD", "DEM", "INDEF", "NN", "NPROP", "POSS", "PPRO", "REL", "V", "WPRO"]
 art_subcats = ["Def", "Indef", "Neg"]
 pro_subcats = ["Pers", "Refl"]
@@ -39,9 +39,13 @@ seg_werden = "werd<~>en"
 
 particle_boundary = "<#>"
 
-index_haben  = None
-index_sein   = 3
-index_werden = None
+lemma_index_haben  = None
+lemma_index_sein   = 3
+lemma_index_werden = None
+
+paradigm_index_haben  = None
+paradigm_index_sein   = None
+paradigm_index_werden = 1
 
 lexcat = ["pos",
           "subcat",
@@ -62,15 +66,21 @@ parcat = ["degree",
 Lexcat = namedtuple("Lexcat", lexcat, defaults=[None] * len(lexcat[1:]))
 Parcat = namedtuple("Parcat", parcat, defaults=[None] * len(parcat))
 
-Formspec  = namedtuple("Formspec",  ["index", "lexcat", "parcat"])
-Lemmaspec = namedtuple("Lemmaspec", ["index", "segmented_lemma", "pos"])
+Formspec  = namedtuple("Formspec",  ["lemma_index", "paradigm_index", "lexcat", "parcat"])
+Lemmaspec = namedtuple("Lemmaspec", ["lemma_index", "paradigm_index", "segmented_lemma", "pos"])
 
 def string(value):
     return str(value or "")
 
-def format_index(index):
-    if index:
-        return "<IDX" + string(index) + ">"
+def format_lemma_index(lemma_index):
+    if lemma_index:
+        return "<IDX" + string(lemma_index) + ">"
+    else:
+        return ""
+
+def format_paradigm_index(paradigm_index):
+    if paradigm_index:
+        return "<PAR" + string(paradigm_index) + ">"
     else:
         return ""
 
@@ -80,76 +90,77 @@ def format_pos(pos):
 def format_categories(categories):
     return "".join(["<" + cat + ">" for cat in categories])
 
-def generate_forms(transducer, index, seg, pos, categories):
-    forms = transducer.generate(seg + format_index(index)
+def generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories):
+    forms = transducer.generate(seg + format_lemma_index(lemma_index)
+                                    + format_paradigm_index(paradigm_index)
                                     + format_pos(pos)
                                     + format_categories(categories))
     return forms
 
-def generate_old_forms(transducer, index, seg, pos, categories):
-    forms = generate_forms(transducer, index, seg, pos, categories + ["Old"])
+def generate_old_forms(transducer, lemma_index, paradigm_index, seg, pos, categories):
+    forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories + ["Old"])
     return forms
 
-def generate_nonstandard_forms(transducer, index, seg, pos, categories):
-    forms = generate_forms(transducer, index, seg, pos, categories + ["NonSt"])
+def generate_nonstandard_forms(transducer, lemma_index, paradigm_index, seg, pos, categories):
+    forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories + ["NonSt"])
     return forms
 
-def add_forms(formdict, index, lexcat, parcat, forms):
+def add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms):
     if forms:
-        formdict[Formspec(index, lexcat, parcat)] = sorted(set(forms))
+        formdict[Formspec(lemma_index, paradigm_index, lexcat, parcat)] = sorted(set(forms))
 
-def add_additional_forms(formdict, index, lexcat, parcat, forms):
-    if (index, lexcat, parcat) in formdict:
-        formdict[Formspec(index, lexcat, parcat)].extend(sorted(set(forms)))
+def add_additional_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms):
+    if (lemma_index, paradigm_index, lexcat, parcat) in formdict:
+        formdict[Formspec(lemma_index, paradigm_index, lexcat, parcat)].extend(sorted(set(forms)))
     else:
-        formdict[Formspec(index, lexcat, parcat)] = sorted(set(forms))
+        formdict[Formspec(lemma_index, paradigm_index, lexcat, parcat)] = sorted(set(forms))
 
-def add_old_forms(formdict, index, lexcat, parcat, forms):
+def add_old_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms):
     if forms:
         formatted_forms = [form + " (va.)" for form in forms]
-        add_additional_forms(formdict, index, lexcat, parcat, formatted_forms)
+        add_additional_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, formatted_forms)
 
-def add_nonstandard_forms(formdict, index, lexcat, parcat, forms):
+def add_nonstandard_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms):
     if forms:
         formatted_forms = [form + " (ugs.)" for form in forms]
-        add_additional_forms(formdict, index, lexcat, parcat, formatted_forms)
+        add_additional_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, formatted_forms)
 
-def add_superlative_forms(formdict, index, lexcat, parcat, forms):
+def add_superlative_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms):
     if forms:
         complex_forms = ["am " + form for form in forms]
-        add_forms(formdict, index, lexcat, parcat, complex_forms)
+        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, complex_forms)
 
-def add_perfect_forms(formdict, index, lexcat, parcat, forms, participle):
+def add_perfect_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms, participle):
     if forms:
         complex_forms = [form + " " + participle for form in forms]
-        add_forms(formdict, index, lexcat, parcat, complex_forms)
+        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, complex_forms)
 
-def add_future_forms(formdict, index, lexcat, parcat, forms, infinitive):
+def add_future_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms, infinitive):
     if forms:
         complex_forms = [form + " " + infinitive for form in forms]
-        add_forms(formdict, index, lexcat, parcat, complex_forms)
+        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, complex_forms)
 
-def add_future_perfect_forms(formdict, index, lexcat, parcat, forms, infinitive, participle):
+def add_future_perfect_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms, infinitive, participle):
     if forms:
         complex_forms = [form + " " + participle + " " + infinitive for form in forms]
-        add_forms(formdict, index, lexcat, parcat, complex_forms)
+        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, complex_forms)
 
-def add_nonfinite_perfect_forms(formdict, index, lexcat, parcat, forms, participle):
+def add_nonfinite_perfect_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms, participle):
     if forms:
         complex_forms = [participle + " " + form for form in forms]
-        add_forms(formdict, index, lexcat, parcat, complex_forms)
+        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, complex_forms)
 
-def add_particle_verb_forms(formdict, index, lexcat, parcat, forms, particle):
+def add_particle_verb_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms, particle):
     if forms:
         complex_forms = [form[len(particle):] + " " + particle for form in forms]
-        add_forms(formdict, index, lexcat, parcat, complex_forms)
+        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, complex_forms)
 
-def add_imperative_particle_verb_forms(formdict, index, lexcat, parcat, forms, particle):
+def add_imperative_particle_verb_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms, particle):
     if forms:
         complex_forms = [form + " " + particle for form in forms]
-        add_forms(formdict, index, lexcat, parcat, complex_forms)
+        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, complex_forms)
 
-def get_noun_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_forms=False):
+def get_noun_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms=False, nonstandard_forms=False):
     formdict = {}
     # nominalised adjectives
     for gender in genders:
@@ -165,8 +176,8 @@ def get_noun_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_
                                   case,
                                   number,
                                   inflection]
-                    forms = generate_forms(transducer, index, seg, pos, categories)
-                    add_forms(formdict, index, lexcat, parcat, forms)
+                    forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                    add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                     if old_forms:
                         # no such forms
                         pass
@@ -184,17 +195,17 @@ def get_noun_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_
                 categories = [gender,
                               case,
                               number]
-                forms = generate_forms(transducer, index, seg, pos, categories)
-                add_forms(formdict, index, lexcat, parcat, forms)
+                forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                 if old_forms:
-                    forms = generate_old_forms(transducer, index, seg, pos, categories)
-                    add_old_forms(formdict, index, lexcat, parcat, forms)
+                    forms = generate_old_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                    add_old_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                 if nonstandard_forms:
-                    forms = generate_nonstandard_forms(transducer, index, seg, pos, categories)
-                    add_nonstandard_forms(formdict, index, lexcat, parcat, forms)
+                    forms = generate_nonstandard_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                    add_nonstandard_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
     return formdict
 
-def get_adjective_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_forms=False):
+def get_adjective_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms=False, nonstandard_forms=False):
     formdict = {}
     lexcat = Lexcat(pos = pos)
     # predicative forms
@@ -202,11 +213,11 @@ def get_adjective_formdict(transducer, index, seg, pos, old_forms=False, nonstan
         parcat = Parcat(degree   = degree,
                         function = "Pred")
         categories = [degree, "Pred"]
-        predicative_forms = generate_forms(transducer, index, seg, pos, categories)
+        predicative_forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
         if degree == "Sup":
-            add_superlative_forms(formdict, index, lexcat, parcat, predicative_forms)
+            add_superlative_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, predicative_forms)
         else:
-            add_forms(formdict, index, lexcat, parcat, predicative_forms)
+            add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, predicative_forms)
         if old_forms:
             # no such forms
             pass
@@ -222,8 +233,8 @@ def get_adjective_formdict(transducer, index, seg, pos, old_forms=False, nonstan
                         inflection = "Invar")
         categories = [degree,
                       "Invar"]
-        forms = generate_forms(transducer, index, seg, pos, categories)
-        add_forms(formdict, index, lexcat, parcat, forms)
+        forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
         if old_forms:
             # no such forms
             pass
@@ -246,8 +257,8 @@ def get_adjective_formdict(transducer, index, seg, pos, old_forms=False, nonstan
                                       case,
                                       number,
                                       inflection]
-                        forms = generate_forms(transducer, index, seg, pos, categories)
-                        add_forms(formdict, index, lexcat, parcat, forms)
+                        forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                         if old_forms:
                             # no such forms
                             pass
@@ -256,7 +267,7 @@ def get_adjective_formdict(transducer, index, seg, pos, old_forms=False, nonstan
                             pass
     return formdict
 
-def get_article_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_forms=False):
+def get_article_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms=False, nonstandard_forms=False):
     formdict = {}
     for subcat in art_subcats:
         lexcat = Lexcat(pos    = pos,
@@ -277,17 +288,17 @@ def get_article_formdict(transducer, index, seg, pos, old_forms=False, nonstanda
                                           case,
                                           number,
                                           inflection]
-                            forms = generate_forms(transducer, index, seg, pos, categories)
-                            add_forms(formdict, index, lexcat, parcat, forms)
+                            forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                            add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                             if old_forms:
                                 # no such forms
                                 pass
                             if nonstandard_forms:
-                                forms = generate_nonstandard_forms(transducer, index, seg, pos, categories)
-                                add_nonstandard_forms(formdict, index, lexcat, parcat, forms)
+                                forms = generate_nonstandard_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                                add_nonstandard_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
     return formdict
 
-def get_cardinal_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_forms=False):
+def get_cardinal_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms=False, nonstandard_forms=False):
     formdict = {}
     lexcat = Lexcat(pos = pos)
     # forms inflected for function, gender, case, number, and inflectional strength
@@ -306,8 +317,8 @@ def get_cardinal_formdict(transducer, index, seg, pos, old_forms=False, nonstand
                                       case,
                                       number,
                                       inflection]
-                        forms = generate_forms(transducer, index, seg, pos, categories)
-                        add_forms(formdict, index, lexcat, parcat, forms)
+                        forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                         if old_forms:
                             # no such forms
                             pass
@@ -316,7 +327,7 @@ def get_cardinal_formdict(transducer, index, seg, pos, old_forms=False, nonstand
                             pass
     return formdict
 
-def get_adjectival_pronoun_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_forms=False):
+def get_adjectival_pronoun_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms=False, nonstandard_forms=False):
     formdict = {}
     lexcat = Lexcat(pos = pos)
     # forms inflected for function, but uninflected for gender, case, number, and inflectional strength
@@ -328,8 +339,8 @@ def get_adjectival_pronoun_formdict(transducer, index, seg, pos, old_forms=False
                         inflection = "Invar")
         categories = [function,
                       "Invar"]
-        forms = generate_forms(transducer, index, seg, pos, categories)
-        add_forms(formdict, index, lexcat, parcat, forms)
+        forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
         if old_forms:
             # no such forms
             pass
@@ -352,17 +363,17 @@ def get_adjectival_pronoun_formdict(transducer, index, seg, pos, old_forms=False
                                       case,
                                       number,
                                       inflection]
-                        forms = generate_forms(transducer, index, seg, pos, categories)
-                        add_forms(formdict, index, lexcat, parcat, forms)
+                        forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                         if old_forms:
                             # no such forms
                             pass
                         if nonstandard_forms:
-                            forms = generate_nonstandard_forms(transducer, index, seg, pos, categories)
-                            add_nonstandard_forms(formdict, index, lexcat, parcat, forms)
+                            forms = generate_nonstandard_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                            add_nonstandard_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
     return formdict
 
-def get_substantival_pronoun_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_forms=False):
+def get_substantival_pronoun_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms=False, nonstandard_forms=False):
     formdict = {}
     # forms with fixed person and inflected for subcat, case, and number
     for person in persons:
@@ -378,14 +389,14 @@ def get_substantival_pronoun_formdict(transducer, index, seg, pos, old_forms=Fal
                                   person,
                                   case,
                                   number]
-                    forms = generate_forms(transducer, index, seg, pos, categories)
-                    add_forms(formdict, index, lexcat, parcat, forms)
+                    forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                    add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                     if old_forms:
-                        forms = generate_old_forms(transducer, index, seg, pos, categories)
-                        add_old_forms(formdict, index, lexcat, parcat, forms)
+                        forms = generate_old_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                        add_old_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                     if nonstandard_forms:
-                        forms = generate_nonstandard_forms(transducer, index, seg, pos, categories)
-                        add_nonstandard_forms(formdict, index, lexcat, parcat, forms)
+                        forms = generate_nonstandard_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                        add_nonstandard_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
     # forms with fixed gender and inflected for subcat, case, and number
     for gender in genders:
         for number in numbers:
@@ -400,17 +411,17 @@ def get_substantival_pronoun_formdict(transducer, index, seg, pos, old_forms=Fal
                                   gender,
                                   case,
                                   number]
-                    forms = generate_forms(transducer, index, seg, pos, categories)
-                    add_forms(formdict, index, lexcat, parcat, forms)
+                    forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                    add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                     if old_forms:
-                        forms = generate_old_forms(transducer, index, seg, pos, categories)
-                        add_old_forms(formdict, index, lexcat, parcat, forms)
+                        forms = generate_old_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                        add_old_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                     if nonstandard_forms:
-                        forms = generate_nonstandard_forms(transducer, index, seg, pos, categories)
-                        add_nonstandard_forms(formdict, index, lexcat, parcat, forms)
+                        forms = generate_nonstandard_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                        add_nonstandard_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
     return formdict
 
-def get_other_pronoun_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_forms=False):
+def get_other_pronoun_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms=False, nonstandard_forms=False):
     formdict = {}
     # forms with fixed gender, but uninflected for case and number
     for gender in genders:
@@ -420,8 +431,8 @@ def get_other_pronoun_formdict(transducer, index, seg, pos, old_forms=False, non
                         number     = "Invar")
         categories = [gender,
                       "Invar"]
-        forms = generate_forms(transducer, index, seg, pos, categories)
-        add_forms(formdict, index, lexcat, parcat, forms)
+        forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
         if old_forms:
             # no such forms
             pass
@@ -439,14 +450,14 @@ def get_other_pronoun_formdict(transducer, index, seg, pos, old_forms=False, non
                 categories = [gender,
                               case,
                               number]
-                forms = generate_forms(transducer, index, seg, pos, categories)
-                add_forms(formdict, index, lexcat, parcat, forms)
+                forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                 if old_forms:
-                    forms = generate_old_forms(transducer, index, seg, pos, categories)
-                    add_old_forms(formdict, index, lexcat, parcat, forms)
+                    forms = generate_old_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                    add_old_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                 if nonstandard_forms:
-                    forms = generate_nonstandard_forms(transducer, index, seg, pos, categories)
-                    add_nonstandard_forms(formdict, index, lexcat, parcat, forms)
+                    forms = generate_nonstandard_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                    add_nonstandard_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
     lexcat = Lexcat(pos = pos)
     # forms inflected for function, but uninflected for gender, case, number, and inflectional strength
     for function in functions:
@@ -457,8 +468,8 @@ def get_other_pronoun_formdict(transducer, index, seg, pos, old_forms=False, non
                         inflection = "Invar")
         categories = [function,
                       "Invar"]
-        forms = generate_forms(transducer, index, seg, pos, categories)
-        add_forms(formdict, index, lexcat, parcat, forms)
+        forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
         if old_forms:
             # no such forms
             pass
@@ -481,17 +492,17 @@ def get_other_pronoun_formdict(transducer, index, seg, pos, old_forms=False, non
                                       case,
                                       number,
                                       inflection]
-                        forms = generate_forms(transducer, index, seg, pos, categories)
-                        add_forms(formdict, index, lexcat, parcat, forms)
+                        forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                        add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                         if old_forms:
                             # no such forms
                             pass
                         if nonstandard_forms:
-                            forms = generate_nonstandard_forms(transducer, index, seg, pos, categories)
-                            add_nonstandard_forms(formdict, index, lexcat, parcat, forms)
+                            forms = generate_nonstandard_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+                            add_nonstandard_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
     return formdict
 
-def get_verb_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_forms=False):
+def get_verb_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms=False, nonstandard_forms=False):
     formdict = {}
     if particle_boundary in seg:
         particle = seg[:seg.find(particle_boundary)]
@@ -501,7 +512,7 @@ def get_verb_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_
         categories = ["PPast",
                       auxiliary]
         # check whether the past participle actually selects the auxiliary
-        past_participle_forms = generate_forms(transducer, index, seg, pos, categories)
+        past_participle_forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
         if past_participle_forms:
             lexcat = Lexcat(pos       = pos,
                             auxiliary = auxiliary)
@@ -509,8 +520,8 @@ def get_verb_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_
             parcat = Parcat(nonfinite = "Inf",
                             tense     = "Pres")
             categories = ["Inf"]
-            infinitive_forms = generate_forms(transducer, index, seg, pos, categories)
-            add_forms(formdict, index, lexcat, parcat, infinitive_forms)
+            infinitive_forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+            add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, infinitive_forms)
             if old_forms:
                 # no such forms
                 pass
@@ -520,7 +531,7 @@ def get_verb_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_
             parcat = Parcat(nonfinite = "Inf",
                             tense     = "Perf")
             for participle in past_participle_forms:
-                add_nonfinite_perfect_forms(formdict, index, lexcat, parcat, [auxiliary], participle)
+                add_nonfinite_perfect_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, [auxiliary], participle)
             if old_forms:
                 # no such forms
                 pass
@@ -531,8 +542,8 @@ def get_verb_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_
             parcat = Parcat(nonfinite = "Part",
                             tense     = "Pres")
             categories = ["PPres"]
-            present_participle_forms = generate_forms(transducer, index, seg, pos, categories)
-            add_forms(formdict, index, lexcat, parcat, present_participle_forms)
+            present_participle_forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
+            add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, present_participle_forms)
             if old_forms:
                 # no such forms
                 pass
@@ -543,7 +554,7 @@ def get_verb_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_
                             tense     = "Perf")
             categories = ["PPast",
                           auxiliary]
-            add_forms(formdict, index, lexcat, parcat, past_participle_forms)
+            add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, past_participle_forms)
             if old_forms:
                 # no such forms
                 pass
@@ -566,10 +577,10 @@ def get_verb_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_
                                                   "Pres",
                                                   mood]
                                     if auxiliary == "haben":
-                                        forms = generate_forms(transducer, index_haben, seg_haben, pos, categories)
+                                        forms = generate_forms(transducer, lemma_index_haben, paradigm_index_haben, seg_haben, pos, categories)
                                     elif auxiliary == "sein":
-                                        forms = generate_forms(transducer, index_sein, seg_sein, pos, categories)
-                                    add_perfect_forms(formdict, index, lexcat, parcat, forms, participle)
+                                        forms = generate_forms(transducer, lemma_index_sein, paradigm_index_sein, seg_sein, pos, categories)
+                                    add_perfect_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms, participle)
                                     if old_forms:
                                         # no such forms
                                         pass
@@ -583,10 +594,10 @@ def get_verb_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_
                                                   "Past",
                                                   mood]
                                     if auxiliary == "haben":
-                                        forms = generate_forms(transducer, index_haben, seg_haben, pos, categories)
+                                        forms = generate_forms(transducer, lemma_index_haben, paradigm_index_haben, seg_haben, pos, categories)
                                     elif auxiliary == "sein":
-                                        forms = generate_forms(transducer, index_sein, seg_sein, pos, categories)
-                                    add_perfect_forms(formdict, index, lexcat, parcat, forms, participle)
+                                        forms = generate_forms(transducer, lemma_index_sein, paradigm_index_sein, seg_sein, pos, categories)
+                                    add_perfect_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms, participle)
                                     if old_forms:
                                         # no such forms
                                         pass
@@ -599,8 +610,8 @@ def get_verb_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_
                                                   number,
                                                   "Pres",
                                                   mood]
-                                    forms = generate_forms(transducer, index_werden, seg_werden, pos, categories)
-                                    add_future_forms(formdict, index, lexcat, parcat, forms, infinitive)
+                                    forms = generate_forms(transducer, lemma_index_werden, paradigm_index_werden, seg_werden, pos, categories)
+                                    add_future_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms, infinitive)
 
                                     if old_forms:
                                         # no such forms
@@ -614,8 +625,8 @@ def get_verb_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_
                                                   number,
                                                   "Pres",
                                                   mood]
-                                    forms = generate_forms(transducer, index_werden, seg_werden, pos, categories)
-                                    add_future_perfect_forms(formdict, index, lexcat, parcat, forms, auxiliary, participle)
+                                    forms = generate_forms(transducer, lemma_index_werden, paradigm_index_werden, seg_werden, pos, categories)
+                                    add_future_perfect_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms, auxiliary, participle)
 
                                     if old_forms:
                                         # no such forms
@@ -628,11 +639,11 @@ def get_verb_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_
                                               number,
                                               tense,
                                               mood]
-                                forms = generate_forms(transducer, index, seg, pos, categories)
+                                forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
                                 if particle:
-                                    add_particle_verb_forms(formdict, index, lexcat, parcat, forms, particle)
+                                    add_particle_verb_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms, particle)
                                 else:
-                                    add_forms(formdict, index, lexcat, parcat, forms)
+                                    add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                                 if old_forms:
                                     # no such forms
                                     pass
@@ -645,11 +656,11 @@ def get_verb_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_
                                 mood   = "Imp")
                 categories = ["Imp",
                               number]
-                forms = generate_forms(transducer, index, seg, pos, categories)
+                forms = generate_forms(transducer, lemma_index, paradigm_index, seg, pos, categories)
                 if particle:
-                    add_particle_verb_forms(formdict, index, lexcat, parcat, forms, particle)
+                    add_particle_verb_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms, particle)
                 else:
-                    add_forms(formdict, index, lexcat, parcat, forms)
+                    add_forms(formdict, lemma_index, paradigm_index, lexcat, parcat, forms)
                 if old_forms:
                     # no such forms
                     pass
@@ -658,37 +669,39 @@ def get_verb_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_
                     pass
     return formdict
 
-def get_formdict(transducer, index, seg, pos, old_forms=False, nonstandard_forms=False):
+def get_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms=False, nonstandard_forms=False):
     # nouns
     if pos == "NN" or pos == "NPROP":
-        formdict = get_noun_formdict(transducer, index, seg, pos, old_forms, nonstandard_forms)
+        formdict = get_noun_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms, nonstandard_forms)
     elif pos == "ADJ":
-        formdict = get_adjective_formdict(transducer, index, seg, pos, old_forms, nonstandard_forms)
+        formdict = get_adjective_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms, nonstandard_forms)
     # articles
     elif pos == "ART":
-        formdict = get_article_formdict(transducer, index, seg, pos, old_forms, nonstandard_forms)
+        formdict = get_article_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms, nonstandard_forms)
     # cardinals
     elif pos == "CARD":
-        formdict = get_cardinal_formdict(transducer, index, seg, pos, old_forms, nonstandard_forms)
+        formdict = get_cardinal_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms, nonstandard_forms)
     # demonstrative and possessive pronouns
     elif pos == "DEM" or pos == "POSS":
-        formdict = get_adjectival_pronoun_formdict(transducer, index, seg, pos, old_forms, nonstandard_forms)
+        formdict = get_adjectival_pronoun_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms, nonstandard_forms)
     # personal pronouns
     elif pos == "PPRO":
-        formdict = get_substantival_pronoun_formdict(transducer, index, seg, pos, old_forms, nonstandard_forms)
+        formdict = get_substantival_pronoun_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms, nonstandard_forms)
     # other pronouns
     elif pos == "INDEF" or pos == "REL" or pos == "WPRO":
-        formdict = get_other_pronoun_formdict(transducer, index, seg, pos, old_forms, nonstandard_forms)
+        formdict = get_other_pronoun_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms, nonstandard_forms)
     # verbs
     elif pos == "V":
-        formdict = get_verb_formdict(transducer, index, seg, pos, old_forms, nonstandard_forms)
+        formdict = get_verb_formdict(transducer, lemma_index, paradigm_index, seg, pos, old_forms, nonstandard_forms)
     return formdict
 
-def paradigm_subset(formdict, index, lexcat):
-    return [(key, value) for key, value in formdict.items() if key.index == index and key.lexcat == lexcat]
+def paradigm_subset(formdict, lemma_index, paradigm_index, lexcat):
+    return [(key, value) for key, value in formdict.items()
+            if key.lemma_index == lemma_index and key.paradigm_index == paradigm_index and key.lexcat == lexcat]
 
 def cat_dict(cats):
-    return {key: value for key, value in cats._asdict().items() if value}
+    return {key: value for key, value in cats._asdict().items()
+            if value}
 
 def cat_list(cats):
     # remove duplicates while preserving order
@@ -697,7 +710,7 @@ def cat_list(cats):
 def output_json(lemma, output, formdict, no_category_names=False, no_lemma=False):
     paradigms = []
     # remove duplicates while preserving order
-    for index, lexcat in list(dict.fromkeys((key.index, key.lexcat) for key in formdict.keys())):
+    for lemma_index, paradigm_index, lexcat in list(dict.fromkeys((key.lemma_index, key.paradigm_index, key.lexcat) for key in formdict.keys())):
         if no_category_names:
             cat_value = cat_list
         else:
@@ -705,14 +718,15 @@ def output_json(lemma, output, formdict, no_category_names=False, no_lemma=False
         if no_lemma:
             paradigms.append({"paradigm": [{**cat_value(key.parcat),
                                             "forms": value}
-                                           for key, value in paradigm_subset(formdict, index, lexcat)]})
+                                           for key, value in paradigm_subset(formdict, lemma_index, paradigm_index, lexcat)]})
         else:
             paradigms.append({"lemma": lemma,
-                              "index": index,
+                              "lemma_index": lemma_index,
+                              "paradigm_index": paradigm_index,
                               **cat_value(lexcat),
                               "paradigm": [{**cat_value(key.parcat),
                                             "forms": value}
-                                           for key, value in paradigm_subset(formdict, index, lexcat)]})
+                                           for key, value in paradigm_subset(formdict, lemma_index, paradigm_index, lexcat)]})
     json.dump(paradigms, output, ensure_ascii=False)
 
 def output_dsv(lemma, output, formdict, no_category_names=False, no_lemma=False, header=True, force_color=False, delimiter="\t"):
@@ -724,7 +738,8 @@ def output_dsv(lemma, output, formdict, no_category_names=False, no_lemma=False,
                                  term.bold("Paradigm Forms")])
         elif no_category_names:
             csv_writer.writerow([term.bold_underline("Lemma"),
-                                 term.underline("Index"),
+                                 term.underline("Lemma Index"),
+                                 term.underline("Paradigm Index"),
                                  term.underline("Categories"),
                                  "Paradigm Categories",
                                  term.bold("Paradigm Forms")])
@@ -742,7 +757,8 @@ def output_dsv(lemma, output, formdict, no_category_names=False, no_lemma=False,
                                  term.bold("Paradigm Forms")])
         else:
             csv_writer.writerow([term.bold_underline("Lemma"),
-                                 term.underline("Index"),
+                                 term.underline("Lemma Index"),
+                                 term.underline("Paradigm Index"),
                                  term.underline("POS"),
                                  term.underline("Subcategory"),
                                  term.underline("Auxiliary"),
@@ -764,7 +780,8 @@ def output_dsv(lemma, output, formdict, no_category_names=False, no_lemma=False,
                                  term.bold(", ".join(formdict[formspec]))])
         elif no_category_names:
             csv_writer.writerow([term.bold_underline(lemma),
-                                 term.underline(string(formspec.index)),
+                                 term.underline(string(formspec.lemma_index)),
+                                 term.underline(string(formspec.paradigm_index)),
                                  term.underline(" ".join(filter(None, formspec.lexcat))),
                                  " ".join(filter(None, formspec.parcat)),
                                  term.bold(", ".join(formdict[formspec]))])
@@ -782,7 +799,8 @@ def output_dsv(lemma, output, formdict, no_category_names=False, no_lemma=False,
                                  term.bold(", ".join(formdict[formspec]))])
         else:
             csv_writer.writerow([term.bold_underline(lemma),
-                                 term.underline(string(formspec.index)),
+                                 term.underline(string(formspec.lemma_index)),
+                                 term.underline(string(formspec.paradigm_index)),
                                  term.underline(formspec.lexcat.pos),
                                  term.underline(string(formspec.lexcat.subcat)),
                                  term.underline(string(formspec.lexcat.auxiliary)),
@@ -799,31 +817,33 @@ def output_dsv(lemma, output, formdict, no_category_names=False, no_lemma=False,
                                  string(formspec.parcat.tense),
                                  term.bold(", ".join(formdict[formspec]))])
 
-def generate_paradigms(transducer, lemma, index=None, pos=None, user_specified=False, old_forms=False, nonstandard_forms=False):
+def generate_paradigms(transducer, lemma, lemma_index=None, paradigm_index=None, pos=None, user_specified=False, old_forms=False, nonstandard_forms=False):
     lemmaspecs = []
     if user_specified:
         if pos in psos:
-            if index in indices:
-                lemmaspecs = [Lemmaspec(index, lemma, pos)]
-            else:
-                lemmaspecs = [Lemmaspec(None, lemma, pos)]
+            lemmaspecs = [Lemmaspec(lemma_index, paradigm_index, lemma, pos)]
     else:
         analyses = analyse_word(transducer, lemma)
-        lemmaspecs = sorted({Lemmaspec(analysis.index, analysis.segmented_lemma, analysis.pos)
+        lemmaspecs = sorted({Lemmaspec(analysis.lemma_index, analysis.paradigm_index, analysis.segmented_lemma, analysis.pos)
                              for analysis in analyses if analysis.lemma == lemma and analysis.pos in psos},
-                            key=lambda l: (l.index, l.segmented_lemma, l.pos))
-        if index in indices:
-            lemmaspecs = [lemmaspec for lemmaspec in lemmaspecs if lemmaspec.index == str(index)]
+                            key=lambda l: (l.lemma_index, l.paradigm_index, l.segmented_lemma, l.pos))
+        if lemma_index in indices:
+            lemmaspecs = [lemmaspec for lemmaspec in lemmaspecs
+                          if lemmaspec.lemma_index == str(lemma_index)]
+        if paradigm_index in indices:
+            lemmaspecs = [lemmaspec for lemmaspec in lemmaspecs
+                          if lemmaspec.paradigm_index == str(paradigm_index)]
         if pos in psos:
-            lemmaspecs = [lemmaspec for lemmaspec in lemmaspecs if lemmaspec.pos == pos]
+            lemmaspecs = [lemmaspec for lemmaspec in lemmaspecs
+                          if lemmaspec.pos == pos]
     formdict = {}
     for lemmaspec in lemmaspecs:
         formdict.update(get_formdict(transducer, *lemmaspec, old_forms=old_forms, nonstandard_forms=nonstandard_forms))
     return formdict
 
-def output_paradigms(transducer, lemma, output, index=None, pos=None, user_specified=False, old_forms=False, nonstandard_forms=False, no_category_names=False, no_lemma=False, header=True, force_color=False, output_format="tsv"):
+def output_paradigms(transducer, lemma, output, lemma_index=None, paradigm_index=None, pos=None, user_specified=False, old_forms=False, nonstandard_forms=False, no_category_names=False, no_lemma=False, header=True, force_color=False, output_format="tsv"):
     term = Terminal(force_styling=force_color)
-    formdict = generate_paradigms(transducer, lemma, index, pos, user_specified, old_forms, nonstandard_forms)
+    formdict = generate_paradigms(transducer, lemma, lemma_index, paradigm_index, pos, user_specified, old_forms, nonstandard_forms)
     if formdict:
         if output_format == "json":
             output_json(lemma, output, formdict, no_category_names, no_lemma)
@@ -832,10 +852,16 @@ def output_paradigms(transducer, lemma, output, index=None, pos=None, user_speci
         else:
             output_dsv(lemma, output, formdict, no_category_names, no_lemma, header, force_color)
     else:
-        if index and pos:
-            print(term.bold(lemma) + ": No such lemma of index {0} and part-of-speech {1}.".format(index, pos), file=sys.stderr)
-        elif index:
-            print(term.bold(lemma) + ": No such lemma of index {0}.".format(index), file=sys.stderr)
+        if lemma_index and paradigm_index and pos:
+            print(term.bold(lemma) + ": No such lemma of lemma index {0}, paradigm index {1}, and part-of-speech {2}.".format(lemma_index, paradigm_index, pos), file=sys.stderr)
+        elif lemma_index and pos:
+            print(term.bold(lemma) + ": No such lemma of lemma index {0} and part-of-speech {1}.".format(lemma_index, pos), file=sys.stderr)
+        elif paradigm_index and pos:
+            print(term.bold(lemma) + ": No such lemma of paradigm index {0} and part-of-speech {1}.".format(paradigm_index, pos), file=sys.stderr)
+        elif lemma_index:
+            print(term.bold(lemma) + ": No such lemma of lemma index {0}.".format(lemma_index), file=sys.stderr)
+        elif paradigm_index:
+            print(term.bold(lemma) + ": No such lemma of paradigm index {0}.".format(paradigm_index), file=sys.stderr)
         elif pos:
             print(term.bold(lemma) + ": No such lemma of part-of-speech {0}.".format(pos), file=sys.stderr)
         elif user_specified:
@@ -857,14 +883,16 @@ def main():
                             help="preserve color and formatting when piping output")
         parser.add_argument("-H", "--no-header", action="store_false",
                             help="suppress table header")
-        parser.add_argument("-i", "--index", type=int, choices=indices,
+        parser.add_argument("-i", "--lemma-index", type=int, choices=indices,
                             help="homographic lemma index")
+        parser.add_argument("-I", "--paradigm-index", type=int, choices=indices,
+                            help="paradigm index")
         parser.add_argument("-j", "--json", action="store_true",
                             help="output JSON object")
         parser.add_argument("-n", "--no-category-names", action="store_true",
                             help="do not output category names")
         parser.add_argument("-N", "--no-lemma", action="store_true",
-                            help="do not output lemma, index, and lexical categories")
+                            help="do not output lemma, lemma index, paradigm index, and lexical categories")
         parser.add_argument("-o", "--old-forms", action="store_true",
                             help="output also archaic forms")
         parser.add_argument("-p", "--pos", choices=psos,
@@ -886,7 +914,7 @@ def main():
             output_format = "csv"
         else:
             output_format = "tsv"
-        output_paradigms(transducer, args.lemma, args.output, args.index, args.pos, args.user_specified, args.old_forms, args.nonstandard_forms, args.no_category_names, args.no_lemma, args.no_header, args.force_color, output_format)
+        output_paradigms(transducer, args.lemma, args.output, args.lemma_index, args.paradigm_index, args.pos, args.user_specified, args.old_forms, args.nonstandard_forms, args.no_category_names, args.no_lemma, args.no_header, args.force_color, output_format)
     except KeyboardInterrupt:
         sys.exit(130)
     # except TypeError: